Подключены ruff и mypy; состояние UI вынесено в отдельный модуль
Статический анализ нашёл два реальных дефекта: assert в
Orchestrator.repository исчезал при запуске с -O, а _num был объявлен
возвращающим str | int вместо int | float. Ещё два поймал mypy —
в build_templates.py фабрика Document стояла в аннотациях вместо типа,
и AppState.start передавал Path | None в Orchestrator.

AppState переехал из main_view.py в app/ui/state.py: там вся логика,
которую можно проверять без запуска Flet, и она закрыта 17 новыми
тестами. В main_view.py осталась только разметка.

PLR намеренно вне набора ruff: на разборе таблиц Word он даёт сотни
срабатываний, ни одно из которых не указывает на дефект.

+++ edu-convert-main/app/core/normalizer.py
@@ -152,7 +152,7 @@ def to_float(value: object) -> float | None:
 def to_int(value: object) -> int | None:
     """Безопасно приводит ячейку к int (``«12»``, ``«12.0»``, ``12.0`` → 12)."""
     number = to_float(value)
-    return int(round(number)) if number is not None else None
+    return round(number) if number is not None else None
 
 
 def as_int(value: object, default: int = 0) -> int:
